backend/requirements.txt
Flask
pdfplumber
google-generativeai
flask-cors